Modify developer

+++ app/Livewire/Post/Edit.php
@@ -107,7 +107,10 @@ class Edit extends ModalComponent
     {
         // delete previous media
         $media = $this->post->media;
-        $this->mediaService->destroy($media);
+
+        if ($media) {
+            $this->mediaService->destroy($media);
+        }
 
         // add updated media
         $url = $this->fileService->storeFile($newMedia);
